CheckedDropdowns
Added checkedDropdowns component for output and python page.
