+++ src/components/pages/python.tsx
@@ -1,6 +1,5 @@
-import { useState, useEffect} from 'react';
 import '../../styles/python.scss';
-import {BooleanDropdown} from '../shared/booleanDropdown';
+import { CheckedDropdowns } from '../shared/checkedDropdowns';
 import {lineOfCode, CodeFormat}  from  '../shared/codeFormat';
 import {Color} from '../shared/constants';
 
@@ -54,25 +53,6 @@ function Python(props:PythonProps): JSX.Element {
       { [')']: Color.White},
     ],
   ];
-
-  const [dropValues, setDropValues] = useState({
-    top: false,
-    middle: false,
-    bottom: false,
-  });
-
-  const selected = (value:string, pos:string) => {
-    const val = value === 'True';
-    const newDropValues = {...dropValues, [pos]: val};
-    setDropValues(newDropValues);
-  };
-
-  useEffect(() => {
-    if (dropValues.top && dropValues.middle && !dropValues.bottom) {
-      props.onCorrect();
-    }
-  },[dropValues]);
-
   return (
     <div className='frame'>
       <div id='python'>
@@ -82,11 +62,7 @@ function Python(props:PythonProps): JSX.Element {
             <CodeFormat code={code}/>
           </div>
           <div id='output'>
-            <div id='dropdowns'>
-              {Object.keys(dropValues).map((dropdown) => (
-                <BooleanDropdown key={dropdown} onChange={(value) => selected(value,dropdown)} position={dropdown}/>
-              ))}
-            </div>
+            <CheckedDropdowns onCorrect={props.onCorrect} topCorrect={true} middleCorrect={true} bottomCorrect={false}/>
           </div>
         </div>
       </div>
