+++ src/styles/checkedDropdowns.scss
@@ -1,0 +1,7 @@
+@import 'global';
+
+#dropdowns {
+  height: 90px;
+  position: absolute;
+  width: 70px;
+}
