+++ src/styles/python.scss
@@ -16,12 +16,6 @@
   color: $main-green;
 }
 
-#dropdowns {
-  height: 90px;
-  position: absolute;
-  width: 70px;
-}
-
 #problem {
   margin-top: 3vw;
 }
